+++ package.json
@@ -45,7 +45,6 @@
     "find-cache-dir": "^3.3.1",
     "glob-parent": "^5.1.1",
     "globby": "^11.0.0",
-    "is-glob": "^4.0.1",
     "loader-utils": "^2.0.0",
     "minimatch": "^3.0.4",
     "normalize-path": "^3.0.0",
